Update importData.py
Casting to np.int in the LHCFillsByNumber

+++ cl2pd/importData.py
@@ -264,7 +264,7 @@ def LHCFillsByNumber(fillList, verbose=False):
 
         if verbose: print(('Fill ' + str(i)))
 
-        DATA=cals.getLHCFillData(i)
+        DATA=cals.getLHCFillData(np.int(i))
 
         fillNumberList=[]
         startTimeList=[]
